HBnB_v2/models/amenity.py
#!/usr/bin/python3
'''
    Implementation of the Amenity class
'''
from models.base_model import BaseModel, Base
import models
import os
from sqlalchemy import Column, String, Integer, ForeignKey, DateTime
from sqlalchemy.orm import relationship


class Amenity(BaseModel, Base):
    '''
        Implementation for the Amenities.
    '''

    __tablename__ = 'amenities'

    if os.getenv('HBNB_TYPE_STORAGE') == 'db':
        from models.place import association_table
        name = Column(String(128), nullable=False)

        place_amenities = relationship("Place", secondary=association_table)
    else:
        name = ""

